+++ app/Http/Controllers/Api/Attachments/AttachmentController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Attachments;
 
 use App\Http\Controllers\Controller;
 use App\Models\Attachment;
