+++ routes/api.php
@@ -1,13 +1,12 @@
 <?php
 
-use App\Http\Controllers\Api\AttachmentController;
-use App\Http\Controllers\Api\AttachmentDownloadController;
+use App\Http\Controllers\Api\Attachments\AttachmentController;
+use App\Http\Controllers\Api\Attachments\DownloadController;
 use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\PasswordController;
 use App\Http\Controllers\Api\ProfileController;
 use App\Http\Controllers\Api\ProfilePictureController;
 use App\Http\Controllers\Api\UserController;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -46,7 +45,7 @@ Route::middleware('auth:sanctum')->group(function () {
         Route::get('/', [AttachmentController::class, 'index'])->name('attachments.index');
         Route::post('/', [AttachmentController::class, 'store'])->name('attachments.store');
         Route::get('/{attachment}', [AttachmentController::class, 'show'])->name('attachments.show');
-        Route::get('/{attachment}/download', AttachmentDownloadController::class)->name('attachments.download');
+        Route::get('/{attachment}/download', DownloadController::class)->name('attachments.download');
         Route::delete('/{attachment}', [AttachmentController::class, 'destroy'])->name('attachments.destroy');
     });
